Fix an energy loss bug in fission step
Fission of Carbon to Helium assumed that helium's power advantage was equal to the power gained by fusing helium, however the products of helium fusion partially skipped carbon. Therefore, carbon fission actually resulted in a large net loss of energy.

This bug has been fixed by using the total difference of power to determine the upwards cost. Note that the power of helium is barely more than carbon. This should probably be adjusted.